Revert "Update module(age gender race recognition model)"
This reverts commit 86d072bee71c33e2155691b04524cb93e1a3e5a5.
