--- Modules/AGR_recognition/efficientnet/__init__.py
@@ -1,10 +1,0 @@
-__version__ = "0.6.1"
-from .model import EfficientNet
-from .utils import (
-    GlobalParams,
-    BlockArgs,
-    BlockDecoder,
-    efficientnet,
-    get_model_params,
-)
-
