--- Modules/AGR_recognition/utils/sync_batchnorm/__init__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-# File   : __init__.py
-# Author : Jiayuan Mao
-# Email  : maojiayuan@gmail.com
-# Date   : 27/01/2018
-#
-# This file is part of Synchronized-BatchNorm-PyTorch.
-# https://github.com/vacancy/Synchronized-BatchNorm-PyTorch
-# Distributed under MIT License.
-
-from .batchnorm import SynchronizedBatchNorm1d, SynchronizedBatchNorm2d, SynchronizedBatchNorm3d
-from .batchnorm import patch_sync_batchnorm, convert_model
-from .replicate import DataParallelWithCallback, patch_replication_callback
