+++ WebAnalyzer/tasks.py
@@ -25,8 +25,8 @@ def module_load_init(**__):
     #   - Add your model
     #   - You can use worker_index if you need to get and set gpu_id
     #       - ex) gpu_id = worker_index % TOTAL_GPU_NUMBER
-    from Modules.AGR_recognition.main import ARG_recognition
-    analyzer = ARG_recognition()
+    from Modules.dummy.main import Dummy
+    analyzer = Dummy()
 
 
 @app.task
