+++ docker-compose-env/mysql.env
@@ -1,4 +1,4 @@
-MYSQL_HOST_ADDR=hidf-engine-agr_mysql
+MYSQL_HOST_ADDR=analysis-engine-mysql
 MYSQL_DATABASE=module_db
 MYSQL_USER=admin
 MYSQL_PASSWORD=password
