+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '2.3'
 
 services:
   main:
-    container_name: hidf-engine-agr_main
+    container_name: analysis-engine-main
     build:
       context: ./
       dockerfile: Dockerfile
@@ -24,13 +24,13 @@ services:
     volumes:
       - ./media:/workspace/media
     ports:
-      - "11000:8000"
-      - "11022:22"
+      - "8001:8000"
+      - "8022:22"
     stdin_open: true
     tty: true
 
   mysql:
-    container_name: hidf-engine-agr_mysql
+    container_name: analysis-engine-mysql
     image: mysql:latest
     restart: always
     command: mysqld --character-set-server=utf8mb4 --collation-server=utf8mb4_unicode_ci
@@ -41,7 +41,7 @@ services:
     volumes:
       - ./mysql-data:/var/lib/mysql
   rabbitmq:
-    container_name: hidf-engine-agr_rabbitmq
+    container_name: analysis-engine-rabbitmq
     image: rabbitmq:3.8.14
     restart: always
     env_file:
