+++ requirements.txt
@@ -10,7 +10,3 @@ mysqlclient==2.0.1
 mysql-connector
 mysql-connector-python
 opencv-python
-torch==1.9.0
-torchvision==0.10.0
-webcolors==1.11.1
-pyyaml
